Add ccSelf support to Endpoint & Client
This patch introduces 3x breaking changes:

- The onMessage method of the RespokeClient.Listener interface now
requires a 5th param, `Boolean didSend`. This param indicates to the
listener whether the message received was *from* the localEndpoint
*sent* to the specified endpoint (didSend=false), or was *from* the
specified endpoint *to* the localEndpoint (didSend=true)
- The onMessage method of the RespokeEndpoint.Listener interface now
requires a 4th param, `Boolean didSend`. This param has the same
meaning as the one on RespokeClient.Listener#onMessage
- The sendMessage method of RespokeEndpoint class now has 4 params
and the 3rd param is now `Boolean ccSelf`. This param indicates
to the Respoke service whether the message should be sent to other
connections of the sender's endpointId or not.