Add Respoke-SDK header to requests
This patch adds the "Respoke-SDK" header to http and ws requests, as
well as the socket.io connect. When the sdk is being used outside of
a released environment (i.e. during development), the version number
of the sdk will not show, because there is no version number to be
displayed in that scenario. When the sdk is used from a distribution
such as Maven, the version number will be displayed along with the
name of the sdk, i.e. "Respoke-Android/1.2.0".

Related to MER-4340.