Start sending iceCandidates as soon as offer sent
In a previous version of Respoke.js (which the android sdk was based
on at the time of writing), we waited for the 'answer' signal to be
received before sending any local iceCandidates to the remote peer.
Now that Respoke interops with Asterisk, it is important that these
iceCandidates be sent as soon as possible after sending the offer,
so that ice buffering will work properly.