Send finalCandidates signal on ice gathering end
As a final improvement to ice candidate sending, this patch implements
the 'finalCandidates' ice candidate signal that was recently added to
the Respoke backend. This signal allows the backend to more accurately
compose an 'offer' signal for clients that do not support trickle ice,
resulting in a better chance for a successful call setup.